Remove constraint on reporting objective
Requiring that the hash accuracy and collisions were larger than some
pre-defined threshold was leftover from when the objective was their
ratio.  With Bhattacharrya distance, there isn't (off the top of my
head) any way to spuriously get a very low value in the same way.  So,
we can remove these constraints to allow solutions where the
distributions may not be close to zero but are very non-overlapping.