Use Cython/__builtin_popcount for int_dist
Using a lookup table was limited in that you had to build tables for
every possible integer size, and what's worse, for bit widths larger
than 16 it took a long time to generate them, they took up a lot of RAM
(too much for 64 bit), and they took much longer to access (because of
paging?).  Using the built-in popcount operation is just as fast (single
operation), but the performance doesn't degrade as the bit width gets
larger and it's O(0) in memory.  Using popcount means using Cython, to
get access to the operation from Python, so we have to split out the
int_dist function to a separate file which is compiled by setup.py.